+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.12] - 2026-09-09
+
+### Changed
+
+- Support integration URL updates and incident recovery controls; reject integration reassignment when updating collections.
+
 ## [0.1.11] - 2026-09-04
 
 ### Added